Handle parameters and data separately.
Add support for passing a perl structure to the do_request() call that will be
converted to JSON and sent in the request body. For POST/PUT calls the
parameters are then put in the query string as required.

Update documentation accordingly with some additional examples.

Joint work with jocar/@theseal.